Compose Engram and shifted hyper connections from shared primitives
